requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.4
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
docarray==0.14.9
docker==5.0.3
fastapi==0.79.0
filelock==3.8.0
fonttools==4.34.4
frozenlist==1.3.1
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
h11==0.13.0
httptools==0.4.0
idna==3.3
jina==3.7.5
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.5.2
multidict==6.0.2
numpy==1.23.1
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
prometheus-client==0.14.1
protobuf==3.20.1
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rich==12.5.1
six==1.16.0
slack-bolt==1.14.3
slack-sdk==3.18.1
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websocket-client==1.3.3
websockets==10.3
yarl==1.8.1
